foerr: remove local Log:: settings
With circular dependency between Log and SettingsManager removed, remove
"cached" settings stored inside Log class and replace them with regular
settings accessed via SettingsManager. This requires tweaking the
initial setup of Log and SettingsManager as they are interconnected in
the short period when both are uninitialized. This is unfortunately the
curse of using static objects, but we only have to do this once at the
very top of main(), so hopefully it won't cause too much problems.

Additionally change some method names to be in line with other classes
(e.g. ::setPosition -> ::handleScreenResize).

Signed-off-by: h67ma <szycikm@gmail.com>