style: Run everything through clang-format
Run clang-format -i on all .hpp and .cpp files and manually fix a few
remaining style issues when necessary. Overall it can be said that the
code was mostly well-formatted before this change.

Most of the changes are related to:
* Include order (now sorted alphabetically, as opposed to RCT, which I
  don't think clang-format supports)
* Wrapping some long lines
* Argument/parameter wrapping and placement
* Lambdas
* Some overlooked minor issues, such as newlines containing only
  whitespace

clang-format and/or its current configuration is not perfect, so there
are some caveats. Most notably, it adds two tabs before struct members,
probably because IndentAccessModifiers = true, and an option
specifically for structs doesn't seem to exist. Sometimes macros are
also weirdly wrapped.

Included are some non-automatic cosmetic changes:
* Removed excessive ',' in DevConsole::commands
* Moved some comments in Room::tick()
* Removed excessive brackets in PipBuckCategoryButton::centerText()
* Added several exclusions for automatic formatting
  (// clang-format off) where it was significantly messing up the code

No functional changes (technically changes in include order might have
broken something, but project still compiles, so it's unlikely).

Signed-off-by: h67ma <szycikm@gmail.com>