foerr: Add New Game page to main menu
Add a simple page to main menu, listing all campaigns in res/campaigns
dir and showing a button for each one. When clicking the buttons,
respective campaign can be loaded.

List of campaigns is currently being loaded only once, in the
constructor. This means that if res/campaigns dir changes on disk during
runtime, there is no way for the changes to be reflected in the page.
In the future the list should auto-refresh after a campaign fails to
load, or upon pressing a manual "Refresh" button on the page.

For now the page uses the default button to display each campaign. In
the future this should be changed to a scrollable list.

Another thing to add would be campaign metadata. Currently each
campaign list item displays campaign ID - it should display campaign
title. It should also display campaign description and other details
on the side, but that should probably be attempted once a scrolling list
is implemented (as it's probable it will include the option to show
selected item detail on the side).

Signed-off-by: h67ma <szycikm@gmail.com>