foerr: Remove Continue button from main menu
Remove the "Continue" button for now, as the savegame system is not yet
implemented, and campaigns can now be loaded via the New Game page.

Don't move main menu buttons up - the page frame was designed to to have
its first item non-selectable, so if we move buttons up now, it would
look weird when New Game page was selected.

Signed-off-by: h67ma <szycikm@gmail.com>