foerr: Make pages begin at (0, 0)
Elements on all GUI pages are currently aligned to where PipBuck starts,
meaning that to define a position of some button/label/etc, every time
some value needs to be added. For example, to add a button at the top
left corner of PipBuck screen, its position would need to be set to
(400, 250), which doesn't make much sense. Additionally, main menu now
also reuses some of the pages, and needs to add some weird negative
offset to each one so it displays in the correct place.

To make things simpler, start defining elements on page starting from
(0, 0) and move the whole page to the correct place.

Signed-off-by: h67ma <szycikm@gmail.com>