foerr: Display log messages in PipBuck Log page
Add a callback to the Log class that is being called when logging a
message. Use it to append a simple (for now) text label on the Log page.
There is a hard arbitrary limit on the amount of last messages, after
reaching it old messages will be removed. Ultimately a scrollable list
should be shown in place of the crude text labels.

Signed-off-by: h67ma <szycikm@gmail.com>