foerr: Add Scrollable
Create base classes needed to build scrollable lists and views.
Scrollable implements the basic ability to scroll drawable content.
ListView makes it possible to create scrollable collections of items of
the same kind. It's possible to find the selected item on click event.

Create the first scrollable container: TextListView, which displays
a collection of text labels. Use it to display log messages on PipBuck
Log page.

Signed-off-by: h67ma <szycikm@gmail.com>